hotfix
distinguishing fatal repo errors

+++ odoo_admin/addons/cicd/models/repository.py
@@ -268,11 +268,16 @@ class Repository(models.Model):
                             try:
                                 shell.checkout_branch(branch)
                             except Exception as ex:
+                                breakpoint()
                                 logsio.error(ex)
-                                logsio.info("Recreating workspace folder")
-                                shell.rm(shell.cwd)
-                                self.clone_repo(machine, shell.cwd, logsio)
-                                shell.checkout_branch(branch)
+
+                                if self._exception_means_broken_repo(ex):
+                                    logsio.info("Recreating workspace folder")
+                                    shell.rm(shell.cwd)
+                                    self.clone_repo(machine, shell.cwd, logsio)
+                                    shell.checkout_branch(branch)
+                                else:
+                                    raise
 
                             if branch in candidate_branch_names:
                                 # avoid fast forward git extra commits
@@ -296,7 +301,7 @@ class Repository(models.Model):
                             shell.X(["git", "submodule", "update", "--init", "--recursive"])
                     except Exception as ex:
                         logger.error('error', exc_info=True)
-                        if '.git/index.lock' in str(ex):
+                        if not self._exception_means_broken_repo(ex):
                             raise RetryableJobError(str(ex), seconds=5, ignore_retry=True) from ex
                         raise
 
@@ -532,3 +537,18 @@ class Repository(models.Model):
                 raise NotImplementedError()
         else:
             raise NotImplementedError()
+
+    def _exception_means_broken_repo(self, ex):
+        """
+        Checks error message from git if the repo is broken and needs recreation.
+        Recreation is a problem, because current state is lost, so this should be 
+        avoided if possible.
+        """
+
+        exmessage = str(ex)
+        if any(x in exmessage for x in [
+            'could not read Username for',
+            '.git/index.lock',
+        ]):
+            return False
+        return True
